E-AdaptiveLearning/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pymongo==4.6.1
typing-extensions==4.9.0
pydantic==2.5.2
langchain==0.0.352
langchain-google-genai==0.1.1
google-generativeai==0.3.1
langgraph==0.0.28
uuid==1.30
python-dateutil==2.8.2